requirements.txt
streamlit==0.63.1
pandas==1.0.5
numpy==1.18.5
scikit-learn==0.23.1
biopython==1.77
matplotlib==3.2.2
neatbio==0.0.2
pycparser==2.20
